requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
bcrypt==4.0.1
PyJWT==2.8.0
